fmt: style nits, rename section, reword intro, add list of formatting traits
Closes #129, cc #196
