--- examples/staging/custom-fmt/input.md
@@ -1,12 +1,0 @@
-The `std::fmt` module defines a number of traits each for one of the formatting
-argument types.
-
-This example demonstrates how to implement the `Show` trait used for
-the default formatting argument:&nbsp;`{}`. Implemention of the other traits
-is exactly the same - the `fmt` function is all that is needed.
-
-See the docs for
-[full list](http://doc.rust-lang.org/std/fmt/index.html#formatting-traits)
-of formatting traits.
-
-{print-custom.play}
